requirements.txt
Flask==2.3.3
bcrypt==4.0.1
Werkzeug==2.3.7
click==8.1.7
colorama==0.4.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
gunicorn==21.2.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
